% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\docType{data}
\name{pm10}
\alias{pm10}
\title{Tibble con i dati di pm10 per 62 stazioni della regione LOMBARDIA}
\format{
Un tibble con 8 colonne e 45322 osservazioni
}
\usage{
pm10
}
\description{
Tibble con i dati di pm10 per 62 stazioni della regione LOMBARDIA
}
\keyword{datasets}
